float & double checked

+++ include/joint.h
@@ -17,54 +17,54 @@ class Joint
 {
  public:
     Joint() {};
-    Joint(Matrix4f, Matrix4f);
+    Joint(Matrix4d, Matrix4d);
     void transform();
     
  protected:
-    Matrix4f parent_to_joint;
-    Matrix4f joint_to_child;
-    float p2j_aa[4];
-    float j2c_aa[4];
-    float p2j_trans[3];
-    float j2c_trans[3];
-    void rotation2angleaxis(Matrix4f, float*);
+    Matrix4d parent_to_joint;
+    Matrix4d joint_to_child;
+    double p2j_aa[4];
+    double j2c_aa[4];
+    double p2j_trans[3];
+    double j2c_trans[3];
+    void rotation2angleaxis(Matrix4d, double*);
     virtual void rotate();
     chrono::system_clock::time_point start_time;
 };
 
 class Revolute: public Joint {
  public:
-    Revolute(Matrix4f, Matrix4f, double, double);
+    Revolute(Matrix4d, Matrix4d, double, double);
     void transform();
     
  private:
     double joint_limit_z[2];
-    float angle_z;
+    double angle_z;
     void rotate();
 };
 
 class BallSocket: public Joint {
  public:
-    BallSocket(Matrix4f, Matrix4f, double, double,
+    BallSocket(Matrix4d, Matrix4d, double, double,
                double, double);
  private:
     double joint_limit_z[2];
     double joint_limit_x[2];
-    float angle_z;
-    float angle_x;
+    double angle_z;
+    double angle_x;
     void rotate();
 };
 
 class Floating: public Joint {
  public:
-    Floating(Matrix4f);
+    Floating(Matrix4d);
     void transform();
     
  private:
-    float x;
-    float y;
-    float z;
-    float angle_x;
-    float angle_y;
-    float angle_z;
+    double x;
+    double y;
+    double z;
+    double angle_x;
+    double angle_y;
+    double angle_z;
 };
